backend/pyproject.toml
[project]
name = "gbase8a-assistant-backend"
version = "0.1.0"
description = "GBase 8a Agent 数据库助手后端"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi>=0.115",
    "greenlet>=3.4.0",
    "litellm>=1.83.4",
    "pydantic>=2.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.3",
    "ruff>=0.15.9",
    "sqlalchemy>=2.0",
    "sqlglot>=30.4.2",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff.format]
quote-style = "double"
